eigen evaluator torch

+++ src/eigensolver/galerkin/constant_basis.py
@@ -10,16 +10,16 @@ class ConstantBasis():
         self.basis_dim = self.base_basis.basis_dim + 1
 
     def __call__(self, x):
-        out = torch.ones([x.size(0),self.basis_dim])
+        out = torch.ones([x.size(0),self.basis_dim], dtype = x.dtype)
         out[:,1:] = self.base_basis(x)
         return out
     
     def grad(self, x):
-        out = torch.zeros([x.size(0),self.basis_dim,self.dim])
+        out = torch.zeros([x.size(0),self.basis_dim,self.dim], dtype = x.dtype)
         out[:,1:,:] = self.base_basis.grad(x)
         return out
     
     def laplacian(self, x):
-        out = torch.zeros([x.size(0),self.basis_dim])
+        out = torch.zeros([x.size(0),self.basis_dim], dtype = x.dtype)
         out[:,1:] = self.base_basis.laplacian(x)
         return out
